Correction de la hiérarchie de certains fichiers, et essai d'ajout des fonctions "virtual" mais sans succès

+++ carre.h
@@ -18,8 +18,8 @@ class carre : public forme{
   carre(point * p, int cote);
 
   //Méthodes
-  virtual float surface();
-  virtual float perimetre();
+  virtual int surface();
+  virtual int perimetre();
   void operator+=(point const & p);
 };
